Crate tmsn

rust_tmsn is a collection of general modules to help implementing TMSN for various learning algorithms.

Use rust_tmsn with Cargo

Please download the source code of rust_tmsn to your computer, and append following lines to the Cargo.toml file in your project with the path should be the actual location of rust_tmsn.

[dependencies]
rust_tmsn = { path = "../rust-tmsn" }

Modules

network
Establish network connections between the workers in the cluster